Move ModuleManifest and WebModuleManifest to its own files
